Fix "bash: have: command not found" error.
It turns out bash-completion defines a function called have(), and then
unsets it at the end.

It might be only on certain versions, not sure -- I only got this error
on my OS X laptop with the bash-completion version homebrew ships (1.3).

After upgrading to the latest bash-completion (2.0), it seemed as though
have() was just available throughout the .bashrc, even after commenting
out the function definition and before sourcing the completion script.
So there's probably more going on.

Let's just pick a different name and avoid this mess.